rerate enhancement

+++ utilities/reusableCompoenents.go
@@ -1657,30 +1657,33 @@ func TDFReraD(iCompany uint, iPolicy uint, iFunction string, iTranno uint) (stri
 		}
 
 	}
-	results = initializers.DB.First(&tdfpolicy, "company_id = ? and policy_id = ? and tdf_type = ?", iCompany, iPolicy, iFunction)
-	if results.Error != nil {
-		tdfpolicy.CompanyID = iCompany
-		tdfpolicy.PolicyID = iPolicy
-		tdfpolicy.Seqno = tdfrule.Seqno
-		tdfpolicy.TDFType = iFunction
-		tdfpolicy.EffectiveDate = oDate
-		tdfpolicy.Tranno = iTranno
-		initializers.DB.Create(&tdfpolicy)
-		return "", nil
-	} else {
-		initializers.DB.Delete(&tdfpolicy)
-		var tdfpolicy models.TDFPolicy
-		tdfpolicy.CompanyID = iCompany
-		tdfpolicy.PolicyID = iPolicy
-		tdfpolicy.Seqno = tdfrule.Seqno
-		tdfpolicy.TDFType = iFunction
-		tdfpolicy.ID = 0
-		tdfpolicy.EffectiveDate = oDate
-		tdfpolicy.Tranno = iTranno
+	if oDate != "" {
+		results = initializers.DB.First(&tdfpolicy, "company_id = ? and policy_id = ? and tdf_type = ?", iCompany, iPolicy, iFunction)
+		if results.Error != nil {
+			tdfpolicy.CompanyID = iCompany
+			tdfpolicy.PolicyID = iPolicy
+			tdfpolicy.Seqno = tdfrule.Seqno
+			tdfpolicy.TDFType = iFunction
+			tdfpolicy.EffectiveDate = oDate
+			tdfpolicy.Tranno = iTranno
+			initializers.DB.Create(&tdfpolicy)
+			return "", nil
+		} else {
+			initializers.DB.Delete(&tdfpolicy)
+			var tdfpolicy models.TDFPolicy
+			tdfpolicy.CompanyID = iCompany
+			tdfpolicy.PolicyID = iPolicy
+			tdfpolicy.Seqno = tdfrule.Seqno
+			tdfpolicy.TDFType = iFunction
+			tdfpolicy.ID = 0
+			tdfpolicy.EffectiveDate = oDate
+			tdfpolicy.Tranno = iTranno
 
-		initializers.DB.Create(&tdfpolicy)
-		return "", nil
+			initializers.DB.Create(&tdfpolicy)
+			return "", nil
+		}
 	}
+	return "", nil
 }
 
 // TDFExpidD - Time Driven Function - Expiry Date Updation
